Saving Accidental Edits
Accidentally made some edits on the SoilWater branch. I am committing the changes just to save them so that I can move the edits to their own respective branches. This commit should not be used for a PR, and this branch is pretty much dead at this point (lots of changes are necessary for the SoilWater class).